# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Room Reservation",
        "description": "API Documentation for Hotel Room Reservation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Tampilkan semua review",
                "description": "Mengambil daftar lengkap semua review dari semua user.",
                "operationId": "ab90a0abe90d35c68285c04be51c5cd6",
                "responses": {
                    "200": {
                        "description": "Daftar review berhasil diambil"
                    },
                    "500": {
                        "description": "Server Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Buat review baru dengan user_id manual",
                "description": "Membuat ulasan baru untuk sebuah reservasi. User ID diinput secara manual.",
                "operationId": "6e21d04d5ceed8dff4d339c3f504e6f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "reservation_id",
                                    "rating"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID User yang memberikan review.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_id": {
                                        "description": "ID Reservasi yang direview.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "description": "Rating dari 1 sampai 5.",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "description": "Komentar atau ulasan.",
                                        "type": "string",
                                        "example": "Sangat memuaskan!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review berhasil dibuat"
                    },
                    "422": {
                        "description": "Validation Error (Data tidak valid)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Tampilkan detail review",
                "description": "Mengambil detail satu data review berdasarkan ID-nya.",
                "operationId": "9d143bc4da3802a9c265f6801f2147f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail review"
                    },
                    "404": {
                        "description": "Not Found (Review tidak ditemukan)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Review"
                ],
                "summary": "Update sebuah review",
                "description": "Memperbarui data review. Idealnya, hanya user yang membuat review atau admin yang bisa melakukan ini.",
                "operationId": "f1b4752d53b5caf380b4825e13be5e8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "description": "Rating dari 1 sampai 5.",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "comment": {
                                        "description": "Komentar atau ulasan.",
                                        "type": "string",
                                        "example": "Cukup baik, tapi bisa lebih bersih."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review berhasil diupdate"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden (jika ada hak akses)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Review"
                ],
                "summary": "Hapus sebuah review",
                "description": "Menghapus review. Idealnya, hanya user yang membuat review atau admin yang bisa melakukan ini.",
                "operationId": "b6e7444309a874f959e6c30d2606d176",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review berhasil dihapus"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden (jika ada hak akses)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}